Move highlight services to their own folder
