+++ src/components/progress-report/highlights/progress-report-highlights.service.ts
@@ -3,11 +3,11 @@ import { UnsecuredDto } from '~/common';
 import {
   withEffectiveSensitivity,
   withScope,
-} from '../authorization/policies/conditions';
-import { Prompt } from '../prompts/dto';
-import { PromptVariantResponseListService } from '../prompts/prompt-variant-response.service';
-import { ProgressReport } from './dto';
-import { ProgressReportHighlight as Highlight } from './dto/hightlights.dto';
+} from '../../authorization/policies/conditions';
+import { Prompt } from '../../prompts/dto';
+import { PromptVariantResponseListService } from '../../prompts/prompt-variant-response.service';
+import { ProgressReport } from '../dto';
+import { ProgressReportHighlight as Highlight } from '../dto/hightlights.dto';
 import { ProgressReportHighlightsRepository } from './progress-report-highlights.repository';
 
 @Injectable()
